Add repo.linaro.org overlay by default
In order to simplify KVM/Xen Jobs, we need repo.linaro.org directly in
images so we don't need to hack them in. With both repo and ppa in
sources.list, we will be able to see what is still in ppa's that needs
moving over to repo - and thus open road to removing ppa's.