alip: switch to chromium
Now that chromium is loaded to linaro-overlay, we can switch over.
For arm64 this is no-brainer, the chromium works a lot better than
iceweasel. For armhf there is not much difference.

The tar.gz grows from 549.2M to 569M, ie 20MB. Some of that is caused
by backported ffmpeg now being part of image.

Change-Id: I83e6a65abd6ead3167bab019e3efd3d52458669b
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>