jessie-arm*-alip: add libxatracker
libxatracker is needed for freedreno. Since we are using mesa from Debian
backport, we need to ensure that this library is installed in the generic
rootfs, as it cannot be pulled properly in the derived images (missing backport
support in hwpack create).

It is likely a temporary change which can be reverted later.

Change-Id: I60b878b8c55462b67656706a79adcd4ef64e9858
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>