jessie-armhf-installer: add new image
This image is intended to be used for the 96boards graphical installer it
contains a minimal graphical system with Xorg using software rasterizer
implementation, based on Debian. The installer is a Qt5 application that runs on
Debian Qt5 w/ X11.

It is an armhf image, such that it can be used on both 32 bit and 64 bit
variants of 96boards.

It includes startup scripts so that the 96boards installer should automatically
startup on boot.

Change-Id: I38727adf73d0a9594eaa49ca470748b9f2f99f72
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>