package-lists: rename *.list.chroot as linaro.list.chroot
There is no specific reason why we are calling the package-list file based on
the image name. All files are already in the appropriate subfolder base on image
name. Having different file names makes it less convenient when trying to look
at differences between images..

This patchset was created using this command:

for i in `ls */customization/package-lists/* `; do \
git mv $i `dirname $i`/linaro.list.chroot; done

And is of course not supposed to bring any change in the generated images.

Change-Id: I93d0d1f8583a9950cbbd9e391346742a2893dbf2
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>